SPI测试：provider与consumer配置需要相同才能通？有问题

+++ example-common/src/main/java/com/goblin/example/common/service/UserService.java
@@ -9,7 +9,7 @@ public interface UserService {
      */
     User getUser(User user);
 
-    default short getNumber(){
-        return 1;
+    default Short getNumber(){
+        return (short) 1;
     }
 }
